+++ src/components/constants.js
@@ -1,0 +1,4 @@
+export const AUTHORS = {
+    ME: 'Me',
+    BOT: 'Bot',
+}
