+++ src/components/message.js
@@ -1,0 +1,21 @@
+import PropTypes from 'prop-types'
+import React from 'react'
+
+function Message(props) {
+    const { nick = '1234id' } = props
+
+    return (
+        <p id={nick}>
+            {props.author}: {props.text}
+        </p>
+    )
+}
+
+Message.propTypes = {
+    nick: PropTypes.string,
+    author: PropTypes.string.isRequired,
+    text: PropTypes.string.isRequired,
+}
+Message.defaultProps = {}
+
+export default Message
